pyproject.toml
[project]
name = "charla_facil"
version = "0.1.0"
description = ""
authors = [{ name = "Michal Bajer", email = "michal.bajer@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = ["google-adk (>=1.18.0,<2.0.0)", "pydantic (>=2.12.4,<3.0.0)", "sqlalchemy (>=2.0.44,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)"]

[tool.poetry]
packages = [{ include = "charla_facil", from = "." }]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
